Separated match from its sqlachemy binding
